%zu not supported by mingw32, so we resort to casting to %lu which is

+++ Rctl/src/correlation.c
@@ -59,7 +59,7 @@ double correlation(const double* x, const double* y, size_t dim, bool verbose){
       XiYi += x[i] * y[i];
       n++;
     }else if(verbose) {
-      info("Missing value at i=%zu\n", i);
+      info("Missing value at i=%lu\n", (unsigned long)i);
     }
   }
   double onedivn = 1.0 / n;
@@ -101,7 +101,7 @@ double* cor1toN(double* x, double** y, size_t dim, size_t ny, int nthreads, bool
         YiP2[j] += y[j][i] * y[j][i];
         n++;
       }else if(verbose) {
-        info("Missing value at i=%zu\n", i);
+        info("Missing value at i=%lu\n", (unsigned long)i);
       }
     }
     onedivn[j] = 1.0 / (double)n;
@@ -117,7 +117,7 @@ double* cor1toN(double* x, double** y, size_t dim, size_t ny, int nthreads, bool
       cors[j] = nom / denom;
     }
     if(isNaN(cors[j]) || isinf(cors[j]) || cors[j] < -(RANGE) || cors[j] > RANGE){ 
-      if(verbose) info("Correlation '%.8f' not in range [-1, 1] [%f %f %zu]\n", cors[j], nom, denom, dim);
+      if(verbose) info("Correlation '%.8f' not in range [-1, 1] [%f %f %lu]\n", cors[j], nom, denom, (unsigned long)dim);
     }
   }
   free(onedivn); free(Yi); free(YiP2); free(XiYi);
@@ -136,7 +136,12 @@ double* getCorrelations(const Phenotypes phenotypes, const Genotypes genotypes, 
       double* P2  = get(phenotypes.data[phe2], inds);
       cors[i]    = correlation(P1, P2, inds.nelements, false);
       if (verbose) {
-        info("CTL: %zu %zu %zu | %d [%zu] -> %f\n", phe1, mar, phe2, genoenc.data[i], inds.nelements, cors[i]);
+        info("CTL: %lu %lu %lu | %d [%lu] -> %f\n", (unsigned long)phe1, 
+                                                    (unsigned long)mar, 
+                                                    (unsigned long)phe2, 
+                                                    genoenc.data[i], 
+                                                    (unsigned long)inds.nelements, 
+                                                    cors[i]);
       }
       free(P1), free(P2); // Clear phenotypes
       free(inds.data);    // Clear index data
@@ -181,4 +186,3 @@ double chiSQtoP(double Cv, int Dof) {
   if (Cv <= 0 || Dof < 1) return 1.0;
   return pchisq(Cv,(double)Dof, 0, 0);
 }
-
